+++ examples/hello-world.rs
@@ -48,25 +48,13 @@ fn main() {
     let (device, queue) =
         block_on(adapter.request_device(&wgpu::DeviceDescriptor::default(), None)).unwrap();
 
-    let desired_present_mode = wgpu::PresentMode::Mailbox;
-    let fallback_present_mode = wgpu::PresentMode::Fifo;
-    let present_mode = if surface
-        .get_supported_modes(&adapter)
-        .iter()
-        .any(|present_mode| *present_mode == desired_present_mode)
-    {
-        desired_present_mode
-    } else {
-        fallback_present_mode
-    };
-
     // Set up swap chain
     let surface_desc = wgpu::SurfaceConfiguration {
         usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
         format: wgpu::TextureFormat::Bgra8UnormSrgb,
         width: size.width as u32,
         height: size.height as u32,
-        present_mode,
+        present_mode: wgpu::PresentMode::Fifo,
     };
 
     surface.configure(&device, &surface_desc);
@@ -134,7 +122,7 @@ fn main() {
                     format: wgpu::TextureFormat::Bgra8UnormSrgb,
                     width: size.width as u32,
                     height: size.height as u32,
-                    present_mode,
+                    present_mode: wgpu::PresentMode::Fifo,
                 };
 
                 surface.configure(&device, &surface_desc);